change pickup base class to interactable
