+++ aimmo-game/simulation/interactables/conditions.py
