+++ aimmo-game/simulation/interactables/effects.py
