--- aimmo-game/simulation/pickups/__init__.py
@@ -1,12 +1,0 @@
-from simulation.pickups.pickup_types import (ALL_PICKUPS, DamageBoostPickup,
-                                             HealthPickup,
-                                             InvulnerabilityPickup,
-                                             serialize_pickups)
-
-__all__ = [
-    'ALL_PICKUPS',
-    'DamageBoostPickup',
-    'InvulnerabilityPickup',
-    'HealthPickup',
-    'serialize_pickups'
-]
